+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Log error object instead of only the status code
+- Throw TooManyRequetsError to Colossus when Crowdin return 429
 
 ## [0.6.0] - 2019-11-25
 ### Added